Restructured the folders to route-groups  so login doesn't have navbar
