+++ app/(main)/favicon.ico
