+++ app/(main)/layout.tsx
@@ -1,8 +1,7 @@
 import type { Metadata } from "next";
 import { plusJakartaSans } from "@/fonts";
-import "./globals.css";
+import "../globals.css";
 import NavBar from "@/components/NavBar";
-import { ChakraProvider } from "@chakra-ui/react";
 import { Providers } from "./providers";
 
 export const metadata: Metadata = {
